course-work/ai/requirements.txt
flask==2.3.3
flask-cors==4.0.0
boto3==1.28.38
torch==2.6.0
git+https://github.com/openai/CLIP.git
Pillow==10.1.0
numpy==1.26.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0